created fetch.py for data request specific code. did some plumbing to remove the legacy etherscan approach from the side bar but left the code in place. updated a bunch of unit tests.
